+++ ext/agoo/graphql.c
@@ -1647,7 +1647,7 @@ gql_dump_hook(agooReq req) {
     if (NULL == (text = agoo_text_prepend(text, buf, cnt))) {
 	agoo_log_cat(&agoo_error_cat, "Failed to allocate memory for a GraphQL dump.");
     }
-    agoo_res_set_message(req->res, text);
+    agoo_res_message_push(req->res, text, true);
 }
 
 gqlField