+++ test/rack_handler_test.rb
@@ -16,7 +16,7 @@ require 'agoo'
 
 class RackHandlerTest < Minitest::Test
   @@server_started = false
-  
+
   class TellMeHandler
     def initialize
     end
@@ -127,13 +127,13 @@ class RackHandlerTest < Minitest::Test
     req['Accept-Encoding'] = '*'
     req['Accept'] = 'application/json'
     req['User-Agent'] = 'Ruby'
-    
+
     res = Net::HTTP.start(uri.hostname, uri.port) { |h|
       h.request(req)
     }
     assert_equal(Net::HTTPNoContent, res.class)
   end
-  
+
   def test_put
     uri = URI('http://localhost:6467/makeme')
     req = Net::HTTP::Put.new(uri)
@@ -142,7 +142,7 @@ class RackHandlerTest < Minitest::Test
     req['Accept'] = 'application/json'
     req['User-Agent'] = 'Ruby'
     req.body = 'hello'
-    
+
     res = Net::HTTP.start(uri.hostname, uri.port) { |h|
       h.request(req)
     }